% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map_dbl}
\alias{map_chr}
\alias{map_lgl}
\alias{map_df}
\title{map}
\usage{
map(.x, .f, ...)

map_dbl(.x, .f, ...)

map_chr(.x, .f, ...)

map_lgl(.x, .f, ...)

map_df(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector}

\item{.f}{A function}

\item{...}{Additional arguments passed on to .f}
}
\description{
This is the parallelised version of map for full details
see \code{\link[purrr]{map}}
}
