% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.R
\name{map2}
\alias{map2}
\alias{map2_dbl}
\alias{map2_chr}
\alias{map2_lgl}
\alias{map2_df}
\title{map2}
\usage{
map2(.x, .y, .f, ...)

map2_dbl(.x, .y, .f, ...)

map2_chr(.x, .y, .f, ...)

map2_lgl(.x, .y, .f, ...)

map2_df(.x, .y, .f, ...)
}
\arguments{
\item{.x}{Vectors of the same length. A vector of length 1 will be recycled.}

\item{.y}{Vectors of the same length. A vector of length 1 will be recycled.}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to .f.}
}
\description{
This is the parallelised version of map2 for full details
see \code{\link[purrr]{map2}}
}
