% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap.R
\name{pmap}
\alias{pmap}
\alias{pmap_dbl}
\alias{pmap_chr}
\alias{pmap_lgl}
\alias{pmap_df}
\title{pmap}
\usage{
pmap(.l, .f, ...)

pmap_dbl(.l, .f, ...)

pmap_chr(.l, .f, ...)

pmap_lgl(.l, .f, ...)

pmap_df(.l, .f, ...)
}
\arguments{
\item{.l}{A list of lists. The length of .l determines the number of arguments that .f will be called with. List names will be used if present}

\item{.f}{A function, formula, or atomic vector.}

\item{...}{Additional arguments passed on to .f.}
}
\description{
This is the parallelised version of pmap for full details
see \code{\link[purrr]{pmap}}
}
