% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{registerCluster}
\alias{registerCluster}
\title{registerCluster}
\usage{
registerCluster(cl = NULL, nblock = NULL)
}
\arguments{
\item{cl}{a cluster created by parallel::makeCluster()}

\item{nblock}{(optional) Number of blocks to carve data into (recommend to set same as ncore)}
}
\description{
This function registers a cluster created by parallel::makeCluster() to be used to parallise the purrr mapping functions
from the parallel package. See \code{\link[parallel]{clusterApply}} for relevent documentation
}
