% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{rerun}
\alias{rerun}
\title{rerun}
\usage{
rerun(.n, expr)
}
\arguments{
\item{.n}{The number of times to evaluate expr}

\item{expr}{An expression to evaluate .n times}
}
\description{
This is the parallelised version of rerun for full details
see \code{\link[purrr]{rerun}}
}
