% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setClusterSeed}
\alias{setClusterSeed}
\title{setClusterSeed}
\usage{
setClusterSeed(cl, seeds)
}
\arguments{
\item{cl}{A cluster to set the seeds across}

\item{seeds}{Either a single value or a vector. If a single value the seeds for each process are automatically generated.}
}
\description{
This function sets the seeds across all your cluster processes to ensure reproducibility.
}
